requirements.txt
deepface
opencv-python
paho-mqtt
python-dotenv
